+++ test/new_relic/dependency_detection_test.rb
@@ -469,6 +469,28 @@ class DependencyDetectionTest < Minitest::Test
     end
   end
 
+  def test_false_bool_to_string_instrumentation_config
+    name = :false_config_second_detect_test
+    executed = false
+
+    dd = DependencyDetection.defer do
+      named(name)
+      depends_on { NewRelic::Agent.config[:"instrumentation.#{name}"] }
+      executes { executed = true }
+    end
+
+    with_config(:"instrumentation.#{name}" => false) do
+      DependencyDetection.detect!
+
+      assert_equal :disabled, dd.config_value
+
+      # the way the bug worked, it would get installed when it shouldn't be but only on the second time detect! was called
+      DependencyDetection.detect!
+
+      refute executed
+    end
+  end
+
   def test_already_executed_items_are_not_executed_again
     unexecuted = Minitest::Mock.new
     unexecuted.expect :executed, false